issue #9 : Slug feature is implemented, sort options fixed

+++ core-application/webapp/mapr-rest/src/main/java/com/mapr/music/service/impl/AlbumServiceImpl.java
@@ -2,18 +2,18 @@ package com.mapr.music.service.impl;
 
 import com.mapr.music.dao.MaprDbDao;
 import com.mapr.music.dao.SortOption;
-import com.mapr.music.dao.impl.AlbumDao;
-import com.mapr.music.dao.impl.ArtistsDao;
 import com.mapr.music.dto.AlbumDto;
 import com.mapr.music.dto.ResourceDto;
 import com.mapr.music.model.Album;
-import com.mapr.music.model.Artist;
 import com.mapr.music.service.AlbumService;
 import com.mapr.music.service.PaginatedService;
 import org.apache.commons.beanutils.PropertyUtilsBean;
 
+import javax.inject.Inject;
 import javax.inject.Named;
+import javax.ws.rs.NotFoundException;
 import java.lang.reflect.InvocationTargetException;
+import java.util.Arrays;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -21,7 +21,6 @@ import java.util.stream.Collectors;
 /**
  * Actual implementation of {@link AlbumService} which is responsible of performing all business logic.
  */
-@Named
 public class AlbumServiceImpl implements AlbumService, PaginatedService {
 
     private static final long ALBUMS_PER_PAGE_DEFAULT = 50;
@@ -33,6 +32,8 @@ public class AlbumServiceImpl implements AlbumService, PaginatedService {
     private static final String[] ALBUM_SHORT_INFO_FIELDS = {
             "_id",
             "name",
+            "slug_name",
+            "slug_postfix",
             "genre",
             "style",
             "barcode",
@@ -44,17 +45,14 @@ public class AlbumServiceImpl implements AlbumService, PaginatedService {
             "artist_list"
     };
 
+
     private final MaprDbDao<Album> albumDao;
-    private final MaprDbDao<Artist> artistDao;
+    private final SlugService slugService;
 
-    public AlbumServiceImpl() {
-        this.albumDao = new AlbumDao();
-        this.artistDao = new ArtistsDao();
-    }
-
-    public AlbumServiceImpl(MaprDbDao<Album> albumDao, MaprDbDao<Artist> artistDao) {
+    @Inject
+    public AlbumServiceImpl(@Named("albumDao") MaprDbDao<Album> albumDao, SlugService slugService) {
         this.albumDao = albumDao;
-        this.artistDao = artistDao;
+        this.slugService = slugService;
     }
 
     @Override
@@ -109,6 +107,20 @@ public class AlbumServiceImpl implements AlbumService, PaginatedService {
      */
     @Override
     public ResourceDto<AlbumDto> getAlbumsPage(Long perPage, Long page, String order, List<String> orderFields) {
+        return getAlbumsPage(perPage, page, Arrays.asList(new SortOption(order, orderFields)));
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @param perPage     specifies number of albums per page. In case when value is <code>null</code> the
+     *                    default value will be used. Default value depends on implementation class.
+     * @param page        specifies number of page, which will be returned. In case when page value is <code>null</code> the
+     *                    first page will be returned.
+     * @param sortOptions sortOptions specifies albums ordering.
+     * @return albums page resource.
+     */
+    public ResourceDto<AlbumDto> getAlbumsPage(Long perPage, Long page, List<SortOption> sortOptions) {
 
         if (page == null) {
             page = FIRST_PAGE_NUM;
@@ -129,12 +141,6 @@ public class AlbumServiceImpl implements AlbumService, PaginatedService {
         ResourceDto<AlbumDto> albumsPage = new ResourceDto<>();
         albumsPage.setPagination(getPaginationInfo(page, perPage));
         long offset = (page - 1) * perPage;
-
-        SortOption[] sortOptions = null;
-        if (order != null && orderFields != null && orderFields.size() > 0) {
-            SortOption.Order sortOrder = SortOption.Order.valueOf(order.toUpperCase());
-            sortOptions = new SortOption[]{new SortOption(sortOrder, orderFields)};
-        }
 
         List<Album> albums = albumDao.getList(offset, perPage, sortOptions, ALBUM_SHORT_INFO_FIELDS);
         List<AlbumDto> albumDtoList = albums.stream()
@@ -159,7 +165,33 @@ public class AlbumServiceImpl implements AlbumService, PaginatedService {
             throw new IllegalArgumentException("Album's identifier can not be empty");
         }
 
-        return albumToDto(albumDao.getById(id));
+        Album album = albumDao.getById(id);
+        if (album == null) {
+            throw new NotFoundException("Album with id '" + id + "' not found");
+        }
+
+        return albumToDto(album);
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @param slugName slug representation of album's name which is used to generate readable and SEO-friendly URLs.
+     * @return album with specified slug name.
+     */
+    @Override
+    public AlbumDto getAlbumBySlugName(String slugName) {
+
+        if (slugName == null || slugName.isEmpty()) {
+            throw new IllegalArgumentException("Album's slug name can not be empty");
+        }
+
+        Album album = slugService.getAlbumBySlug(slugName);
+        if (album == null) {
+            throw new NotFoundException("Album with slug name '" + slugName + "' not found");
+        }
+
+        return albumToDto(album);
     }
 
     /**
@@ -172,6 +204,10 @@ public class AlbumServiceImpl implements AlbumService, PaginatedService {
 
         if (id == null || id.isEmpty()) {
             throw new IllegalArgumentException("Album's identifier can not be empty");
+        }
+
+        if (!albumDao.exists(id)) {
+            throw new NotFoundException("Album with id '" + id + "' not found");
         }
 
         albumDao.deleteById(id);
@@ -193,8 +229,11 @@ public class AlbumServiceImpl implements AlbumService, PaginatedService {
         String id = UUID.randomUUID().toString();
         album.setId(id);
 
+        slugService.setSlugForAlbum(album);
+
         return albumToDto(albumDao.create(album));
     }
+
 
     /**
      * {@inheritDoc}
@@ -205,12 +244,7 @@ public class AlbumServiceImpl implements AlbumService, PaginatedService {
      */
     @Override
     public AlbumDto updateAlbum(Album album) {
-
-        if (album == null || album.getId() == null || album.getId().isEmpty()) {
-            throw new IllegalArgumentException("Album's identifier can not be empty");
-        }
-
-        return albumToDto(albumDao.update(album.getId(), album));
+        return updateAlbum(album.getId(), album);
     }
 
     /**
@@ -231,10 +265,15 @@ public class AlbumServiceImpl implements AlbumService, PaginatedService {
             throw new IllegalArgumentException("Album can not be null");
         }
 
+        if (!albumDao.exists(id)) {
+            throw new NotFoundException("Album with id '" + id + "' not found");
+        }
+
         return albumToDto(albumDao.update(id, album));
     }
 
     private AlbumDto albumToDto(Album album) {
+
         AlbumDto albumDto = new AlbumDto();
         PropertyUtilsBean propertyUtilsBean = new PropertyUtilsBean();
         try {
@@ -242,6 +281,9 @@ public class AlbumServiceImpl implements AlbumService, PaginatedService {
         } catch (IllegalAccessException | InvocationTargetException | NoSuchMethodException e) {
             throw new RuntimeException("Can not create album Data Transfer Object", e);
         }
+
+        String slug = slugService.getSlugForAlbum(album);
+        albumDto.setSlug(slug);
 
         return albumDto;
     }
